+++ com.albumai.pro/tests/run-alb130-tests.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+"use strict";
+
+require("./alb130-v1.2.0-release-candidate.test");