+++ com.albumai.pro/tests/run-alb131-tests.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+"use strict";
+
+require("./alb131-v1.2.0-publication-readiness.test");